Move utils using listers to plugin
This is done to eliminate the dependency from pkg/utils/gardener
to the generate clientsets under pkg/client.
Generated informers/listers are only used in admission plugins,
hence plugin/utils is a good fit for these helper functions.
