+++ plugin/pkg/plant/admission.go
@@ -25,7 +25,6 @@ import (
 	admissioninitializer "github.com/gardener/gardener/pkg/apiserver/admission/initializer"
 	gardencoreinformers "github.com/gardener/gardener/pkg/client/core/informers/internalversion"
 	gardencorelisters "github.com/gardener/gardener/pkg/client/core/listers/core/internalversion"
-	gutil "github.com/gardener/gardener/pkg/utils/gardener"
 	admissionutils "github.com/gardener/gardener/plugin/pkg/utils"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -174,7 +173,7 @@ func (a *Handler) validate(plant *core.Plant, attrs admission.Attributes) error 
 		return err
 	}
 
-	project, err := gutil.ProjectForNamespaceFromInternalLister(a.projectLister, plant.Namespace)
+	project, err := admissionutils.ProjectForNamespaceFromInternalLister(a.projectLister, plant.Namespace)
 	if err != nil {
 		return apierrors.NewBadRequest(fmt.Sprintf("could not find referenced project: %+v", err.Error()))
 	}
